rename: post -> article 으로 변경
- post 메서드와 혼동을 방지하기 위해 이름 변경하였음
