+++ src/main/java/codesquad/springcafe/database/user/UserDatabase.java
@@ -1,4 +1,4 @@
-package codesquad.springcafe.database;
+package codesquad.springcafe.database.user;
 
 import codesquad.springcafe.model.User;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.Optional;
 public interface UserDatabase {
     User save(User user);
 
-    Optional<User> findByNickname(String nickname);
+    Optional<User> findBy(String nickname);
 
     List<User> findAll();
