+++ src/main/resources/static/css/article/show.css
