+++ src/main/resources/templates/article/form.html
@@ -1,39 +1,25 @@
 <!DOCTYPE html>
-<html lang="kr">
+<html lang="kr" xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta http-equiv="content-type" content="text/html; charset=UTF-8">
     <meta charset="utf-8">
     <title>kakao Java Web Programming</title>
     <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-    <link href="../../static/css/reset.css" rel="stylesheet">
+    <link th:href="@{/css/reset.css}" href="../../static/css/reset.css" rel="stylesheet">
     <!--[if lt IE 9]>
     <script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
     <![endif]-->
-    <link href="../../static/css/post/form.css" rel="stylesheet">
+    <link th:href="@{/css/article/form.css}" href="../../static/css/article/form.css" rel="stylesheet">
 </head>
 <body>
 <div class="background">
-    <nav class="navbar navbar-fixed-top header">
-        <div class="col-md-12">
-            <div class="navbar-header">
-                <a href="../index.html" class="navbar-brand">
-                    <img src="../../static/images/KakaoCI.png"/>
-                </a>
-
-            </div>
-            <ul class="nav navbar-nav navbar-right">
-                <li class="active"><a href="../user/list.html">멤버 리스트</a></li>
-                <li><a class="black-component" href="../user/login.html" role="button">로그인</a></li>
-                <li><a class="black-component" href="../user/form.html" role="button">회원가입</a></li>
-            </ul>
-        </div>
-    </nav>
+    <div th:replace="~{fragments/navbar :: navbar}"></div>
 
     <div class="container" id="main">
         <header class="post-header">
             <p class="title">글쓰기</p>
         </header>
-        <form class="submit-write" method="get" action="../index.html ">
+        <form class="submit-write" method="post" action>
             <div class="text-box">
                 <div class="comment-writer-container">
                     <p class="post-title">제목</p>
