+++ src/main/resources/templates/article/show.html
@@ -9,7 +9,7 @@
     <!--[if lt IE 9]>
     <script src="//html5shim.googlecode.com/svn/trunk/html5.js"></script>
     <![endif]-->
-    <link href="../../static/css/post/show.css" rel="stylesheet">
+    <link href="../../static/css/article/show.css" rel="stylesheet">
 </head>
 <body>
     <div class="background">
